Move log to internal
