+++ cmd/tool/slowest/ast.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 
 	"golang.org/x/tools/go/packages"
-	"gotest.tools/gotestsum/log"
+	"gotest.tools/gotestsum/internal/log"
 	"gotest.tools/gotestsum/testjson"
 )
